LayoutTests/http/tests/storageAccess/request-and-grant-storage-access-cross-origin-sandboxed-iframe-expected.txt

CONFIRM: Do you want to use your  ID on ?
CONSOLE MESSAGE: line 64: PASS document.hasStorageAccess was granted.
Tests that cross-origin iframe storage access is granted if the iframe is sandboxed, has the allow token, and the user opts in.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
